inimigo inteligente, um monte de coisa pq tava esquecendo de usar o git

+++ src/GameObject.java
@@ -2,11 +2,11 @@ import java.awt.*;
 
 public abstract class GameObject {
 
-    protected int x,y;
+    protected float x,y;
     protected Identity id;
-    protected int velX, velY;
+    protected float velX, velY;
 
-    public GameObject(int x, int y, Identity id){
+    public GameObject(float x, float y, Identity id){
         this.x = x;
         this.y = y;
         this.id = id;
@@ -16,19 +16,19 @@ public abstract class GameObject {
     public abstract void render(Graphics g);
     public abstract Rectangle getBounds();
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
@@ -40,19 +40,19 @@ public abstract class GameObject {
         this.id = id;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
-    public void setVelX(int velX) {
+    public void setVelX(float velX) {
         this.velX = velX;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
 
-    public void setVelY(int velY) {
+    public void setVelY(float velY) {
         this.velY = velY;
     }
 }
