+++ src/HUD.java
@@ -2,7 +2,7 @@ import java.awt.*;
 
 public class HUD {
 
-    protected static int HEALTH = 100;
+    protected static float HEALTH = 100;
     private int green = 255;
     private  int score = 0;
     private int level = 1;
@@ -10,14 +10,14 @@ public class HUD {
 
     public void tick(){
         HEALTH = Game.clamp(HEALTH, 0, 100);
-        green = HEALTH*2;
+        green = (int)HEALTH*2;
 
     }
     public void render(Graphics g){
         g.setColor(Color.darkGray);
         g.fillRect(15,15,200,20);
         g.setColor(new Color(75, green, 0));
-        g.fillRect(15,15,HEALTH * 2,20);
+        g.fillRect(15,15,(int)HEALTH * 2,20);
         g.setColor(Color.white);
         g.drawString("Score"+score,15,50);
         g.drawString("Level"+level,15,70);
